requirements.txt
amqp==5.3.1
asgiref==3.9.2
attrs==25.4.0
autobahn==25.9.1
Automat==25.4.16
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clickhouse-connect==0.9.2
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.6
django-bootstrap5==25.2
django-debug-toolbar==6.0.0
django-filter==25.1
django-jinja==2.11.0
django-otp==1.6.3
djangorestframework==3.16.1
djangorestframework-xml==2.0.0
et_xmlfile==2.0.0
humanize==4.13.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
kombu==5.5.4
lz4==4.4.4
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyMuPDF==1.26.4
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qrcode==8.2
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
xlsxwriter==3.2.9
zope.interface==8.0.1
zstandard==0.25.0
